Mimictalk_Talking_System/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
openai>=1.0.0
ffmpeg-python